requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
openai==1.3.5
PyPDF2==3.0.1
python-docx==0.8.11
numpy==1.25.2
requests==2.31.0 